requirements.txt
SQLAlchemy==2.0.27
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-dotenv==1.0.1
pytesseract==0.3.10
pdf2image==1.17.0
pandas==2.2.0
alembic==1.13.1
click==8.1.7
tqdm==4.66.1
aiohttp==3.9.3
python-dateutil==2.8.2 